test(map): split remaining oversized map test files under 500-line guideline
Splits the two map test files left over the 500-line guideline by PR #3606
into focused per-concern files, with no behavioural change (every test is
preserved, only regrouped):

- init_game_map_view_builder_test_part2_test.dart (947 lines) -> 4 files
  (port co-location, visibility/unit-presence, civilian markers, warp markers).
- tile_map_visualization_test.dart (610 lines) -> render + helpers test files,
  sharing topology/result fixtures via a new non-test fixtures helper.

All split files are now well under 500 lines. Full colonizethis_map suite
passes (340 tests), preserving determinism.

Refs #3588
