+++ packages/colonizethis_map/test/init_game_map_view_builder_port_colocation_test.dart
@@ -1,0 +1,173 @@
+import 'package:colonizethis_test/test.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_map/colonizethis_map.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+void main() {
+  group('buildInitGameMapViewData port/town co-location', () {
+    test(
+      'town markers: co-located port and town shifts port drawable to N sea cell',
+      () {
+        final owMap = TileMapResult(
+          width: 2,
+          height: 2,
+          grid: [
+            ['p1', 's1'],
+            ['p1', 'p1'],
+          ],
+        );
+        final nwMap = TileMapResult(
+          width: 1,
+          height: 1,
+          grid: [
+            ['p1'],
+          ],
+        );
+        final owTopology = MapTopology(
+          nodes: const [
+            TopologyNode(
+              id: 'p1',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.province,
+            ),
+            TopologyNode(
+              id: 's1',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.seaZone,
+            ),
+          ],
+          edges: const [TopologyEdge(id1: 'p1', id2: 's1')],
+        );
+        final nwTopology = MapTopology(
+          nodes: const [
+            TopologyNode(
+              id: 'p1',
+              regionId: 'newWorld',
+              type: TopologyNodeType.province,
+            ),
+          ],
+          edges: const [],
+        );
+        final game = Game(
+          id: 'townPortColoc',
+          worldState: WorldState(
+            turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
+            oldWorld: RegionData(
+              provinces: [
+                Province(
+                  id: 'oldWorld|p1',
+                  regionId: 'oldWorld',
+                  townTileKey: 'oldWorld|p1|1|1',
+                ),
+              ],
+              units: const [],
+            ),
+            newWorld: const RegionData(provinces: [], units: []),
+            portsByProvinceSeaboard: {
+              'oldWorld|p1|seaboard': 'oldWorld|p1|1|1',
+            },
+          ),
+          players: const [],
+          minorNations: const [],
+          tribes: const [],
+        );
+
+        final viewData = buildInitGameMapViewData(
+          game: game,
+          tileMapByRegion: {'oldWorld': owMap, 'newWorld': nwMap},
+          topologyByRegion: {'oldWorld': owTopology, 'newWorld': nwTopology},
+          cellSize: 8,
+        );
+
+        final tm = viewData.oldWorld.townMarkers.single;
+        expect(tm.isPort, isTrue);
+        expect(tm.x, 1);
+        expect(tm.y, 1);
+        expect(tm.portIconX, 1);
+        expect(tm.portIconY, 0);
+      },
+    );
+    test(
+      'town markers: isPort from seaboard key when value province segment mismatches',
+      () {
+        final owMap = TileMapResult(
+          width: 3,
+          height: 2,
+          grid: [
+            ['p2', 'p2', 'p2'],
+            ['p2', 'p2', 's1'],
+          ],
+        );
+        final nwMap = TileMapResult(
+          width: 1,
+          height: 1,
+          grid: [
+            ['p1'],
+          ],
+        );
+        final owTopology = MapTopology(
+          nodes: const [
+            TopologyNode(
+              id: 'p2',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.province,
+            ),
+            TopologyNode(
+              id: 's1',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.seaZone,
+            ),
+          ],
+          edges: const [TopologyEdge(id1: 'p2', id2: 's1')],
+        );
+        final nwTopology = MapTopology(
+          nodes: const [
+            TopologyNode(
+              id: 'p1',
+              regionId: 'newWorld',
+              type: TopologyNodeType.province,
+            ),
+          ],
+          edges: const [],
+        );
+        final game = Game(
+          id: 'nonCapitalPortKey',
+          worldState: WorldState(
+            turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
+            oldWorld: RegionData(
+              provinces: [
+                const Province(
+                  id: 'oldWorld|p2',
+                  regionId: 'oldWorld',
+                  townTileKey: 'oldWorld|p2|0|0',
+                ),
+              ],
+              units: const [],
+            ),
+            newWorld: const RegionData(provinces: [], units: []),
+            portsByProvinceSeaboard: {
+              'oldWorld|p2|sb': 'oldWorld|p1|2|0',
+            },
+          ),
+          players: const [],
+          minorNations: const [],
+          tribes: const [],
+        );
+
+        final viewData = buildInitGameMapViewData(
+          game: game,
+          tileMapByRegion: {'oldWorld': owMap, 'newWorld': nwMap},
+          topologyByRegion: {'oldWorld': owTopology, 'newWorld': nwTopology},
+          cellSize: 8,
+        );
+
+        final tm = viewData.oldWorld.townMarkers.single;
+        expect(tm.provinceId, 'p2');
+        expect(tm.isPort, isTrue);
+        expect(tm.portIconX, 2);
+        expect(tm.portIconY, 1);
+        expect(viewData.oldWorld.portMarkers.single.provinceId, 'p2');
+      },
+    );
+  });
+}
