+++ packages/colonizethis_map/test/tile_map_visualization_render_test.dart
@@ -1,9 +1,10 @@
-import 'dart:io';
 import 'package:colonizethis_test/test.dart';
 
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_map/colonizethis_map.dart';
 import 'package:image/image.dart' as img;
+
+import 'tile_map_visualization_test_fixtures.dart';
 
 /// Sea color and plains color from tile_map_visualization (for pixel assertions).
 const (int, int, int) _seaRgb = (20, 60, 140);
@@ -12,88 +13,23 @@ const (int, int, int) _seaZoneBorderRgb = (173, 216, 230);
 const int _colorTolerance = 2;
 
 void main() {
-  final topology = MapTopology(
-    nodes: [
-      const TopologyNode(
-        id: 'p1',
-        regionId: 'oldWorld',
-        type: TopologyNodeType.province,
-      ),
-      const TopologyNode(
-        id: 'p2',
-        regionId: 'oldWorld',
-        type: TopologyNodeType.province,
-      ),
-      const TopologyNode(
-        id: 's1',
-        regionId: 'oldWorld',
-        type: TopologyNodeType.seaZone,
-      ),
-    ],
-    edges: [
-      const TopologyEdge(id1: 'p1', id2: 'p2'),
-      const TopologyEdge(id1: 'p1', id2: 's1'),
-    ],
-  );
-
-  final smallResult = TileMapResult(
-    width: 4,
-    height: 3,
-    grid: [
-      ['p1', 'p1', 'p2', 'p2'],
-      ['p1', 's1', 's1', 'p2'],
-      ['p1', 'p1', 'p2', 'p2'],
-    ],
-  );
-
-  /// Result with terrain: p1/p2 land, s1 sea; at least one horizontal and one vertical border.
-  final resultWithTerrain = TileMapResult(
-    width: 4,
-    height: 3,
-    grid: [
-      ['p1', 'p1', 'p2', 'p2'],
-      ['p1', 's1', 's1', 'p2'],
-      ['p1', 'p1', 'p2', 'p2'],
-    ],
-    terrainGrid: [
-      [
-        TerrainType.plains,
-        TerrainType.plains,
-        TerrainType.hills,
-        TerrainType.hills,
-      ],
-      [TerrainType.plains, null, null, TerrainType.hills],
-      [
-        TerrainType.plains,
-        TerrainType.plains,
-        TerrainType.hills,
-        TerrainType.hills,
-      ],
-    ],
-  );
-
-  /// Like resultWithTerrain with resourceGrid: (0,0)=grain, (2,0)=timber, (0,2)=iron; others null.
-  final resultWithTerrainAndResources = TileMapResult(
-    width: 4,
-    height: 3,
-    grid: resultWithTerrain.grid,
-    terrainGrid: resultWithTerrain.terrainGrid,
-    resourceGrid: [
-      [Resource.grain, null, Resource.timber, null],
-      [null, null, null, null],
-      [Resource.iron, null, null, null],
-    ],
-  );
-
   group('renderTileMapToPng', () {
     test('returns non-empty PNG bytes', () {
-      final bytes = renderTileMapToPng(smallResult, topology, cellSize: 4);
+      final bytes = renderTileMapToPng(
+        visualizationSmallResult,
+        visualizationTopology,
+        cellSize: 4,
+      );
       expect(bytes, isNotEmpty);
       expect(bytes.length, greaterThan(100));
     });
 
     test('decoded image has expected dimensions (map + legend)', () {
-      final bytes = renderTileMapToPng(smallResult, topology, cellSize: 8);
+      final bytes = renderTileMapToPng(
+        visualizationSmallResult,
+        visualizationTopology,
+        cellSize: 8,
+      );
       final decoded = img.decodeImage(bytes);
       expect(decoded, isNotNull);
       // Map: 4*8 x 3*8 = 32 x 24. Legend below.
@@ -106,14 +42,14 @@ void main() {
       () {
         const cellSize = 8;
         final bytes = renderTileMapToPng(
-          resultWithTerrain,
-          topology,
+          visualizationResultWithTerrain,
+          visualizationTopology,
           cellSize: cellSize,
         );
         final decoded = img.decodeImage(bytes);
         expect(decoded, isNotNull);
-        final mapW = resultWithTerrain.width * cellSize;
-        final mapH = resultWithTerrain.height * cellSize;
+        final mapW = visualizationResultWithTerrain.width * cellSize;
+        final mapH = visualizationResultWithTerrain.height * cellSize;
         const legendPadding = 12;
         const legendLineHeight = 20;
         const titleLines = 2;
@@ -128,8 +64,8 @@ void main() {
       // Use cellSize 24 so region id label (top-left) does not overlap cell center.
       const cellSize = 24;
       final bytes = renderTileMapToPng(
-        resultWithTerrain,
-        topology,
+        visualizationResultWithTerrain,
+        visualizationTopology,
         cellSize: cellSize,
       );
       final decoded = img.decodeImage(bytes);
@@ -146,8 +82,8 @@ void main() {
     test('land terrain color: center pixel of plains cell matches palette', () {
       const cellSize = 8;
       final bytes = renderTileMapToPng(
-        resultWithTerrain,
-        topology,
+        visualizationResultWithTerrain,
+        visualizationTopology,
         cellSize: cellSize,
       );
       final decoded = img.decodeImage(bytes);
@@ -173,8 +109,8 @@ void main() {
     test('border pixel is black between different region ids (land border)', () {
       const cellSize = 8;
       final bytes = renderTileMapToPng(
-        resultWithTerrain,
-        topology,
+        visualizationResultWithTerrain,
+        visualizationTopology,
         cellSize: cellSize,
       );
       final decoded = img.decodeImage(bytes);
@@ -242,13 +178,13 @@ void main() {
       () {
         const cellSize = 8;
         final bytes = renderTileMapToPng(
-          resultWithTerrain,
-          topology,
+          visualizationResultWithTerrain,
+          visualizationTopology,
           cellSize: cellSize,
         );
         final decoded = img.decodeImage(bytes);
         expect(decoded, isNotNull);
-        final mapH = resultWithTerrain.height * cellSize;
+        final mapH = visualizationResultWithTerrain.height * cellSize;
         const legendPadding = 12;
         const legendLineHeight = 20;
         final legendLines = 2 + 1 + TerrainType.values.length;
@@ -258,7 +194,11 @@ void main() {
     );
 
     test('fallback no terrain: valid PNG, region dimensions, no throw', () {
-      final bytes = renderTileMapToPng(smallResult, topology, cellSize: 8);
+      final bytes = renderTileMapToPng(
+        visualizationSmallResult,
+        visualizationTopology,
+        cellSize: 8,
+      );
       expect(bytes, isNotEmpty);
       final decoded = img.decodeImage(bytes);
       expect(decoded, isNotNull);
@@ -271,8 +211,8 @@ void main() {
       () {
         const cellSize = 8;
         final bytes = renderTileMapToPng(
-          smallResult,
-          topology,
+          visualizationSmallResult,
+          visualizationTopology,
           cellSize: cellSize,
           landSeedPositions: [(0, 0), (1, 1)],
         );
@@ -291,13 +231,13 @@ void main() {
       () {
         const cellSize = 8;
         final bytesWithout = renderTileMapToPng(
-          smallResult,
-          topology,
+          visualizationSmallResult,
+          visualizationTopology,
           cellSize: cellSize,
         );
         final bytesWith = renderTileMapToPng(
-          smallResult,
-          topology,
+          visualizationSmallResult,
+          visualizationTopology,
           cellSize: cellSize,
           landSeedPositions: [(0, 0), (1, 1)],
         );
@@ -331,14 +271,14 @@ void main() {
       () {
         const cellSize = 8;
         final bytesWithContinent = renderTileMapToPng(
-          resultWithTerrain,
-          topology,
+          visualizationResultWithTerrain,
+          visualizationTopology,
           cellSize: cellSize,
           continentSeedPositions: [(0, 0)],
         );
         final bytesWithout = renderTileMapToPng(
-          resultWithTerrain,
-          topology,
+          visualizationResultWithTerrain,
+          visualizationTopology,
           cellSize: cellSize,
         );
         final decodedWith = img.decodeImage(bytesWithContinent);
@@ -358,8 +298,8 @@ void main() {
       () {
         const cellSize = 8;
         final bytesWithBoth = renderTileMapToPng(
-          resultWithTerrain,
-          topology,
+          visualizationResultWithTerrain,
+          visualizationTopology,
           cellSize: cellSize,
           landSeedPositions: [(2, 2)],
           continentSeedPositions: [(0, 0)],
@@ -367,8 +307,8 @@ void main() {
         final decoded = img.decodeImage(bytesWithBoth);
         expect(decoded, isNotNull);
         final bytesLandOnly = renderTileMapToPng(
-          resultWithTerrain,
-          topology,
+          visualizationResultWithTerrain,
+          visualizationTopology,
           cellSize: cellSize,
           landSeedPositions: [(2, 2)],
         );
@@ -400,8 +340,8 @@ void main() {
       () {
         const cellSize = 24;
         final bytes = renderTileMapToPng(
-          resultWithTerrain,
-          topology,
+          visualizationResultWithTerrain,
+          visualizationTopology,
           cellSize: cellSize,
           landSeedPositions: [(0, 0), (2, 2)],
           landSeedContinentIndices: [0, 1],
@@ -411,8 +351,8 @@ void main() {
         // Region id labels are drawn on top of seeds, so we do not assert exact marker pixel colors.
         // Assert that per-continent legend adds two rows (Continent 0, Continent 1).
         final bytesSingleRow = renderTileMapToPng(
-          resultWithTerrain,
-          topology,
+          visualizationResultWithTerrain,
+          visualizationTopology,
           cellSize: cellSize,
           landSeedPositions: [(0, 0), (2, 2)],
         );
@@ -430,13 +370,13 @@ void main() {
       () {
         const cellSize = 8;
         final bytesWithout = renderTileMapToPng(
-          resultWithTerrain,
-          topology,
+          visualizationResultWithTerrain,
+          visualizationTopology,
           cellSize: cellSize,
         );
         final bytesWith = renderTileMapToPng(
-          resultWithTerrainAndResources,
-          topology,
+          visualizationResultWithTerrainAndResources,
+          visualizationTopology,
           cellSize: cellSize,
         );
         final decodedWithout = img.decodeImage(bytesWithout);
@@ -481,8 +421,8 @@ void main() {
       const cellSize = 8;
       const idInset = 2;
       final bytes = renderTileMapToPng(
-        smallResult,
-        topology,
+        visualizationSmallResult,
+        visualizationTopology,
         cellSize: cellSize,
       );
       final decoded = img.decodeImage(bytes);
@@ -500,111 +440,6 @@ void main() {
         isTrue,
         reason: 'Region id label (red) should appear in top-left of first cell',
       );
-    });
-  });
-
-  group('writeTileMapImageToTempFile', () {
-    test('returns path and file exists with content', () {
-      final path = writeTileMapImageToTempFile(smallResult, topology);
-      expect(path, isNotEmpty);
-      final file = File(path);
-      expect(file.existsSync(), isTrue);
-      expect(file.lengthSync(), greaterThan(100));
-    });
-  });
-
-  group('geographicGameWorldResourceGlyphLetter', () {
-    test('returns letters only for grain, timber, iron resource ids', () {
-      expect(geographicGameWorldResourceGlyphLetter('grain'), 'g');
-      expect(geographicGameWorldResourceGlyphLetter('timber'), 't');
-      expect(geographicGameWorldResourceGlyphLetter('iron'), 'i');
-    });
-
-    test(
-      'returns null for ids outside geographic subset, empty, or invalid',
-      () {
-        expect(geographicGameWorldResourceGlyphLetter('gold'), isNull);
-        expect(geographicGameWorldResourceGlyphLetter('notAResource'), isNull);
-        expect(geographicGameWorldResourceGlyphLetter(null), isNull);
-        expect(geographicGameWorldResourceGlyphLetter(''), isNull);
-      },
-    );
-  });
-
-  group('geographicGameWorldLegendResources', () {
-    test('is grain, timber, iron per SPEC geographic legend scope', () {
-      expect(geographicGameWorldLegendResources, const [
-        Resource.grain,
-        Resource.timber,
-        Resource.iron,
-      ]);
-    });
-  });
-
-  group('resource glyph helper iterables', () {
-    test('tileMapResourceGlyphs yields only non-null resources', () {
-      final result = TileMapResult(
-        width: 2,
-        height: 2,
-        grid: [
-          ['p1', 'p1'],
-          ['p1', 'p1'],
-        ],
-        resourceGrid: [
-          [Resource.grain, null],
-          [Resource.iron, null],
-        ],
-      );
-      final glyphs = tileMapResourceGlyphs(result).toList();
-      expect(glyphs, hasLength(2));
-      expect(glyphs[0], (x: 0, y: 0, letter: 'g'));
-      expect(glyphs[1], (x: 0, y: 1, letter: 'i'));
-    });
-
-    test(
-      'geographicGameWorldResourceGlyphs yields only subset resource ids',
-      () {
-        final cells = [
-          const CellViewData(
-            x: 0,
-            y: 0,
-            regionCellId: 'p1',
-            isSea: false,
-            resourceId: 'grain',
-          ),
-          const CellViewData(
-            x: 1,
-            y: 0,
-            regionCellId: 'p1',
-            isSea: false,
-            resourceId: 'gold',
-          ),
-          const CellViewData(
-            x: 0,
-            y: 1,
-            regionCellId: 'p1',
-            isSea: false,
-            resourceId: null,
-          ),
-        ];
-        final glyphs = geographicGameWorldResourceGlyphs(cells).toList();
-        expect(glyphs, hasLength(1));
-        expect(glyphs.single, (x: 0, y: 0, letter: 'g'));
-      },
-    );
-  });
-
-  group('openInDefaultViewer', () {
-    test('does not throw; returns bool', () {
-      final path = writeTileMapImageToTempFile(smallResult, topology);
-      final result = openInDefaultViewer(path);
-      expect(result, isA<bool>());
-    });
-
-    test('returns false when SUPPRESS_IMAGE_VIEWER is 1', () {
-      if (Platform.environment['SUPPRESS_IMAGE_VIEWER'] == '1') {
-        expect(openInDefaultViewer('dummy_path'), isFalse);
-      }
     });
   });
 }
